+++ crates/cranelift/Cargo.toml
@@ -11,13 +11,17 @@ keywords = ["webassembly", "wasm"]
 edition = "2018"
 
 [dependencies]
+anyhow = "1.0"
 wasmtime-environ = { path = "../environ", version = "0.29.0" }
 cranelift-wasm = { path = "../../cranelift/wasm", version = "0.76.0" }
 cranelift-codegen = { path = "../../cranelift/codegen", version = "0.76.0" }
 cranelift-frontend = { path = "../../cranelift/frontend", version = "0.76.0" }
 cranelift-entity = { path = "../../cranelift/entity", version = "0.76.0" }
+cranelift-native = { path = '../../cranelift/native', version = '0.76.0' }
+wasmtime-debug = { path = '../debug', version = '0.29.0' }
 wasmparser = "0.80.0"
 target-lexicon = "0.12"
+gimli = "0.25.0"
 
 [features]
 all-arch = ["cranelift-codegen/all-arch"]
