+++ crates/debug/src/write_debuginfo.rs
@@ -6,25 +6,7 @@ use wasmtime_environ::ir::Endianness;
 use wasmtime_environ::isa::{unwind::UnwindInfo, TargetIsa};
 use wasmtime_environ::{CompiledFunctions, DebugInfoData, ModuleMemoryOffset};
 
-#[derive(Clone)]
-pub enum DwarfSectionRelocTarget {
-    Func(usize),
-    Section(&'static str),
-}
-
-#[derive(Clone)]
-pub struct DwarfSectionReloc {
-    pub target: DwarfSectionRelocTarget,
-    pub offset: u32,
-    pub addend: i32,
-    pub size: u8,
-}
-
-pub struct DwarfSection {
-    pub name: &'static str,
-    pub body: Vec<u8>,
-    pub relocs: Vec<DwarfSectionReloc>,
-}
+pub use wasmtime_environ::{DwarfSection, DwarfSectionReloc, DwarfSectionRelocTarget};
 
 fn emit_dwarf_sections(
     isa: &dyn TargetIsa,
