+++ crates/environ/src/compilation.rs
@@ -2,10 +2,14 @@
 //! module.
 
 use crate::{FunctionAddressMap, FunctionBodyData, ModuleTranslation, Tunables, TypeTables};
-use cranelift_codegen::{binemit, ir, isa, isa::unwind::UnwindInfo};
+use anyhow::Result;
+use cranelift_codegen::{binemit, ir, isa::unwind::UnwindInfo};
 use cranelift_entity::PrimaryMap;
 use cranelift_wasm::{DefinedFuncIndex, FuncIndex, WasmError, WasmFuncType};
 use serde::{Deserialize, Serialize};
+use std::borrow::Cow;
+use std::collections::HashMap;
+use std::fmt;
 use thiserror::Error;
 
 #[allow(missing_docs)]
@@ -93,37 +97,173 @@ pub enum CompileError {
     DebugInfoNotSupported,
 }
 
-/// An implementation of a compiler from parsed WebAssembly module to native
-/// code.
+/// Abstract trait representing the ability to create a `Compiler` below.
+///
+/// This is used in Wasmtime to separate compiler implementations, currently
+/// mostly used to separate Cranelift from Wasmtime itself.
+pub trait CompilerBuilder: Send + Sync + fmt::Debug {
+    /// Like the `Clone` trait, but for the boxed trait object.
+    fn clone(&self) -> Box<dyn CompilerBuilder>;
+
+    /// Sets the target of compilation to the target specified.
+    fn target(&mut self, target: target_lexicon::Triple) -> Result<()>;
+
+    /// Returns the currently configured target triple that compilation will
+    /// produce artifacts for.
+    fn triple(&self) -> &target_lexicon::Triple;
+
+    /// Compiler-specific method to configure various settings in the compiler
+    /// itself.
+    ///
+    /// This is expected to be defined per-compiler. Compilers should return
+    /// errors for unknown names/values.
+    fn set(&mut self, name: &str, val: &str) -> Result<()>;
+
+    /// Compiler-specific method for configuring settings.
+    ///
+    /// Same as [`CompilerBuilder::set`] except for enabling boolean flags.
+    /// Currently cranelift uses this to sometimes enable a family of settings.
+    fn enable(&mut self, name: &str) -> Result<()>;
+
+    /// Returns a list of all possible settings that can be configured with
+    /// [`CompilerBuilder::set`] and [`CompilerBuilder::enable`].
+    fn settings(&self) -> Vec<Setting>;
+
+    /// Builds a new [`Compiler`] object from this configuration.
+    fn build(&self) -> Box<dyn Compiler>;
+}
+
+/// Description of compiler settings returned by [`CompilerBuilder::settings`].
+#[derive(Clone, Copy, Debug)]
+pub struct Setting {
+    /// The name of the setting.
+    pub name: &'static str,
+    /// The description of the setting.
+    pub description: &'static str,
+    /// The kind of the setting.
+    pub kind: SettingKind,
+    /// The supported values of the setting (for enum values).
+    pub values: Option<&'static [&'static str]>,
+}
+
+/// Different kinds of [`Setting`] values that can be configured in a
+/// [`CompilerBuilder`]
+#[derive(Clone, Copy, Debug)]
+pub enum SettingKind {
+    /// The setting is an enumeration, meaning it's one of a set of values.
+    Enum,
+    /// The setting is a number.
+    Num,
+    /// The setting is a boolean.
+    Bool,
+    /// The setting is a preset.
+    Preset,
+}
+
+/// An implementation of a compiler which can compile WebAssembly functions to
+/// machine code and perform other miscellaneous tasks needed by the JIT runtime.
 pub trait Compiler: Send + Sync {
-    /// Compile a function with the given `TargetIsa`.
+    /// Compiles the function `index` within `translation`.
+    ///
+    /// The body of the function is available in `data` and configuration
+    /// values are also passed in via `tunables`. Type information in
+    /// `translation` is all relative to `types`.
     fn compile_function(
         &self,
         translation: &ModuleTranslation<'_>,
         index: DefinedFuncIndex,
         data: FunctionBodyData<'_>,
-        isa: &dyn isa::TargetIsa,
         tunables: &Tunables,
         types: &TypeTables,
     ) -> Result<CompiledFunction, CompileError>;
 
-    /// Creates a trampoline which the host can use to enter wasm. The
-    /// trampoline has type `VMTrampoline` and will call a function of type `ty`
-    /// specified.
-    fn host_to_wasm_trampoline(
-        &self,
-        isa: &dyn isa::TargetIsa,
-        ty: &WasmFuncType,
-    ) -> Result<CompiledFunction, CompileError>;
+    /// Creates a trampoline which the host can use to enter wasm.
+    ///
+    /// The generated trampoline will have the type `VMTrampoline` and will
+    /// call a function of type `ty` specified.
+    fn host_to_wasm_trampoline(&self, ty: &WasmFuncType) -> Result<CompiledFunction, CompileError>;
 
     /// Creates a trampoline suitable for a wasm module to import.
     ///
     /// The trampoline has the type specified by `ty` and will call the function
-    /// `host_fn` which has type `VMTrampoline`.
+    /// `host_fn` which has type `VMTrampoline`. Note that `host_fn` is
+    /// directly embedded into the generated code so this is not suitable for a
+    /// cached value or if `host_fn` does not live as long as the compiled
+    /// function.
+    ///
+    /// This is primarily used for `Func::new` in `wasmtime`.
     fn wasm_to_host_trampoline(
         &self,
-        isa: &dyn isa::TargetIsa,
         ty: &WasmFuncType,
         host_fn: usize,
     ) -> Result<CompiledFunction, CompileError>;
+
+    /// Creates DWARF debugging data for a compilation unit.
+    ///
+    /// This function maps DWARF information found in a wasm module to native
+    /// DWARF debugging information. This is currently implemented by the
+    /// `wasmtime-debug` crate.
+    fn emit_dwarf(
+        &self,
+        debuginfo_data: &crate::DebugInfoData,
+        funcs: &CompiledFunctions,
+        memory_offset: &crate::ModuleMemoryOffset,
+    ) -> Result<Vec<DwarfSection>>;
+
+    /// Returns the target triple that this compiler is compiling for.
+    fn triple(&self) -> &target_lexicon::Triple;
+
+    /// If supported by the target creates a SystemV CIE used for dwarf
+    /// unwinding information.
+    fn create_systemv_cie(&self) -> Option<gimli::write::CommonInformationEntry>;
+
+    /// Returns a list of configured settings for this compiler.
+    fn flags(&self) -> HashMap<String, FlagValue>;
+
+    /// Same as [`Compiler::flags`], but ISA-specific (a cranelift-ism)
+    fn isa_flags(&self) -> HashMap<String, FlagValue>;
+}
+
+#[allow(missing_docs)]
+pub struct DwarfSection {
+    pub name: &'static str,
+    pub body: Vec<u8>,
+    pub relocs: Vec<DwarfSectionReloc>,
+}
+
+#[allow(missing_docs)]
+#[derive(Clone)]
+pub struct DwarfSectionReloc {
+    pub target: DwarfSectionRelocTarget,
+    pub offset: u32,
+    pub addend: i32,
+    pub size: u8,
+}
+
+#[allow(missing_docs)]
+#[derive(Clone)]
+pub enum DwarfSectionRelocTarget {
+    Func(usize),
+    Section(&'static str),
+}
+
+/// Value of a configured setting for a [`Compiler`]
+#[derive(Serialize, Deserialize, Hash, Eq, PartialEq)]
+pub enum FlagValue {
+    /// Name of the value that has been configured for this setting.
+    Enum(Cow<'static, str>),
+    /// The numerical value of the configured settings.
+    Num(u8),
+    /// Whether the setting is on or off.
+    Bool(bool),
+}
+
+impl fmt::Display for FlagValue {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::Enum(v) => v.fmt(f),
+            Self::Num(v) => v.fmt(f),
+            Self::Bool(v) => v.fmt(f),
+        }
+    }
 }
