+++ crates/environ/src/data_structures.rs
@@ -9,12 +9,6 @@ pub mod ir {
     pub use cranelift_codegen::{ValueLabelsRanges, ValueLocRange};
 }
 
-pub mod settings {
-    pub use cranelift_codegen::settings::{
-        builder, Builder, Configurable, Flags, OptLevel, SetError, Setting, SettingKind, Value,
-    };
-}
-
 pub mod isa {
     pub use cranelift_codegen::isa::{
         lookup, unwind, Builder, CallConv, RegUnit, TargetFrontendConfig, TargetIsa,
