+++ crates/jit/Cargo.toml
@@ -15,7 +15,6 @@ wasmtime-environ = { path = "../environ", version = "0.29.0" }
 wasmtime-runtime = { path = "../runtime", version = "0.29.0" }
 wasmtime-cranelift = { path = "../cranelift", version = "0.29.0" }
 wasmtime-lightbeam = { path = "../lightbeam/wasmtime", version = "0.29.0", optional = true }
-wasmtime-debug = { path = "../debug", version = "0.29.0" }
 wasmtime-profiling = { path = "../profiling", version = "0.29.0" }
 wasmtime-obj = { path = "../obj", version = "0.29.0" }
 rayon = { version = "1.0", optional = true }
