+++ crates/lightbeam/wasmtime/Cargo.toml
@@ -12,6 +12,9 @@ readme = "README.md"
 edition = "2018"
 
 [dependencies]
+anyhow = "1.0"
+target-lexicon = "0.12"
+gimli = "0.25"
 lightbeam = { path = "..", version = "0.29.0" }
 wasmparser = "0.80"
 cranelift-codegen = { path = "../../../cranelift/codegen", version = "0.76.0" }
